requirements.txt
# requirements.txt
# MinIO and Object Storage
minio>=7.1.0
urllib3>=1.26.0

# Google Cloud and Vertex AI
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-aiplatform>=1.25.0
google-api-core>=2.11.0
vertexai>=1.25.0

# LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-vertexai>=0.0.1
langchain-experimental>=0.0.50

# Vector database
pgvector>=0.2.0
psycopg2-binary>=2.9.5

# Text processing
regex>=2023.6.3

# Utility packages
uuid>=1.30
json5>=0.9.0
python-dotenv>=1.0.0

# Optional: For enhanced text processing
nltk>=3.8
spacy>=3.6.0
transformers>=4.30.0

# Optional: For PDF processing if needed
PyPDF2>=3.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11
python-multipart>=0.0.6

# Optional: For data validation
pydantic>=2.0.0

# Optional: For logging and monitoring
structlog>=23.1.0
rich>=13.4.0

# Development and testing (optional)
pytest>=7.4.0
asyncio>=3.4.3
asyncpg>=0.28.0
numpy>=1.24.0
pandas>=2.0.0

# Google Cloud and Vertex AI
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-cloud-aiplatform>=1.25.0
google-api-core>=2.11.0
vertexai>=1.25.0

# fastapi
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5